Refactor(woorim/boards): screens 파일들 네비게이션별로 폴더 분류
